included custom usleep

+++ submit/philo/includes/libphilo.h
@@ -6,7 +6,7 @@
 /*   By: jkhong <jkhong@student.42kl.edu.my>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/16 17:51:50 by jkhong            #+#    #+#             */
-/*   Updated: 2021/08/21 18:32:28 by jkhong           ###   ########.fr       */
+/*   Updated: 2021/08/21 19:33:14 by jkhong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,26 +33,26 @@ typedef struct s_globals {
 }				t_globals;
 
 typedef struct s_philo {
-	int				philo_num;
-	unsigned long	last_eat_time;
-	int				times_eaten;
-	int				fork_one;
-	int				fork_two;
+	int		philo_num;
+	size_t	last_eat_time;
+	int		times_eaten;
+	int		fork_one;
+	int		fork_two;
 }				t_philo;
 
-void			initialise_philo(int p_num, t_philo **philo);
-void			initialise_mutex(int p_num, pthread_mutex_t **forks);
-void			create_thread(int p_num, pthread_t **thread,
-					t_philo *philo, void *(f)(void *));
-void			join_and_free_th(int p_num, pthread_t *th_cycle,
-					pthread_t *th_death);
-void			free_mutex_fork_philo(int p_num,
-					pthread_mutex_t *forks, t_philo *philo);
-unsigned long	givetime(void);
-bool			initialise_globals(int params, char *args[],
-					t_globals	*g_args);
-void			end_cycle(t_globals *g_args);
+void	initialise_philo(int p_num, t_philo **philo);
+void	initialise_mutex(int p_num, pthread_mutex_t **forks);
+void	create_thread(int p_num, pthread_t **thread,
+			t_philo *philo, void *(f)(void *));
+void	join_and_free_th(int p_num, pthread_t *th_cycle,
+			pthread_t *th_death);
+void	free_mutex_fork_philo(int p_num,
+			pthread_mutex_t *forks, t_philo *philo);
+size_t	givetime(void);
+bool	initialise_globals(int params, char *args[],
+			t_globals	*g_args);
+void	end_cycle(t_globals *g_args);
 
-void	        ft_usleep(unsigned int n);
+void	ft_usleep(unsigned int n);
 
 #endif
